pyproject.toml
[tool.poetry]
name = "ps-api"
version = "0.1.0"
description = "Project S. API"
authors = ["Simple2B <info.simple2b@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.3"
httpx = "^0.27.0"
uvicorn = "^0.30.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
pymongo = "^4.7.3"
passlib = "^1.7.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"
mypy = "^1.10.0"
pytest = "^8.2.2"
types-python-jose = "^3.3.4.20240106"
mongomock = "^4.1.2"
setuptools = "^70.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "^mypy_cache/",
    "^.pytest_cache/",
    "^.ruff_cache/",
    "^.venv/",
    "^.vscode/",
    "^__pypackages__/",
]
follow_imports = "skip"

[tool.ruff]
exclude = [
    "__pypackages__",
    "__pycache__",
    ".eggs",
    ".git",
    ".github",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
]
line-length = 120
indent-width = 4
